Upgrade to new template from architecture-group. See changelog version 4.0
