--- properties/tomcat/context.xml.default
@@ -1,3 +1,0 @@
-<Context>
-  <Parameter name="licensemodule.log4jFileName" value="${user.home}/services/conf/licensemodule.log4j.xml" override="false" descr="log4j configuration file for licensemodule.war"/>
-</Context>
