--- properties/tomcat/licensemodule.log4j.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-<log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/">
-  <appender name="R" class="org.apache.log4j.RollingFileAppender"> 
-    <param name="file" value="${catalina.base}/logs/licensemodule.log"/>
-    <param name="MaxFileSize" value="10000KB"/>
-    <!-- Keep one backup file -->
-    <param name="MaxBackupIndex" value="10"/>
-    <layout class="org.apache.log4j.PatternLayout"> 
-      <param name="ConversionPattern" value="%d [%t] %-5p %c - %m%n"/> 
-    </layout> 
-  </appender> 
-  <root> 
-    <priority value ="info" /> 
-    <appender-ref ref="R" /> 
-  </root>
-</log4j:configuration>
