+++ src/main/java/dk/statsbiblioteket/doms/licensemodule/service/LicenseModuleApplication.java
@@ -1,0 +1,18 @@
+package dk.statsbiblioteket.doms.licensemodule.service;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+
+import javax.ws.rs.core.Application;
+
+import com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider;
+
+public class LicenseModuleApplication extends Application {
+
+    public Set<Class<?>> getClasses() {
+        return new HashSet<>(Arrays.asList(JacksonJsonProvider.class, LicenseModuleResource.class));
+    }
+
+
+}
