--- src/main/resources/build.properties
@@ -1,5 +1,0 @@
-#$Id: build.properties 1 2011-10-21 09:33:25Z teg $
-# Build Time Information
-APPLICATION.NAME=${pom.name}
-APPLICATION.VERSION=${pom.version}
-APPLICATION.BUILDTIME=${build.time}
