--- src/main/resources/log4j.properties
@@ -1,10 +1,0 @@
-log4j.rootLogger=debug, R
-log4j.appender.C=org.apache.log4j.ConsoleAppender
-log4j.appender.C.layout=org.apache.log4j.PatternLayout
-log4j.appender.C.layout.ConversionPattern=%d [%t] %-5p %c - %m%n
-log4j.appender.R=org.apache.log4j.RollingFileAppender
-log4j.appender.R.File=${catalina.home}/logs/${pom.name}.log
-log4j.appender.R.MaxFileSize=32MB
-log4j.appender.R.MaxBackupIndex=10
-log4j.appender.R.layout=org.apache.log4j.PatternLayout
-log4j.appender.R.layout.ConversionPattern=%d [%t] %-5p %c - %m%n
