+++ src/main/resources/logback.xml
@@ -1,0 +1,6 @@
+<configuration scan="true" scanPeriod="5 minutes">
+    <insertFromJNDI env-entry-name="java:comp/env/licensemodule-logback-config" as="logbackConfiguration"/>
+
+    <include file="${logbackConfiguration}"/>
+
+</configuration>
