+++ src/test/jetty/logback.xml
@@ -1,0 +1,15 @@
+<included>
+    <contextName>java-webapp</contextName>
+
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{35} - %msg %n</pattern>
+        </encoder>
+    </appender>
+
+    <root level="INFO">
+      <appender-ref ref="STDOUT" />
+    </root>
+    <logger name="dk.kb" level="DEBUG" />
+ 
+</included>
